Move classes_addon_rangy_remove_all to its own file
