+++ src/editor_apply_styling_functions.py
@@ -168,20 +168,3 @@ P: I'm losing focus when loading rangy: I use  focusField(0); in my "$(document)
             editor.web.eval("classes_addon_wrap_helper();")
             break
 Editor.my_apply_span_class = my_apply_span_class
-
-
-def classes_addon_rangy_remove_all(editor):
-    # this only works on stuff rangy has highlighted before: so it doesn't help in the browser.
-    js = """
-Object.values(dict).forEach(function (item, index) {
-    item.removeAllHighlights();
-    console.log(item);
-});
-"""
-    # TODO
-    #js = """classes_addon__remove_classes_from_selection();"""
-    #editor.web.eval(js)
-    
-    # at least I can undo the following (which is arguably more important than keeping other formatting)
-    text = editor.web.selectedText()
-    editor.web.eval("setFormat('inserthtml', %s);" % json.dumps(text))
